Use CSharp language version 12 (#13)
Update to CSharp version 12.

Doing this reveals that newly-enforced validation for tracking lifetimes
doesn't allow passing passing the PerfEventBytes struct by `in` ref
without marking the caller as `unsafe`. We don't want the caller to have
to be unsafe, so fix. I'm not entirely sure why this is disallowed, but
the `in` was just a minor performance optimization, so just remove it.