PerfFieldFormat - ignore const, volatile, attribute tokens
PerfFieldFormat should ignore `__attribute__`, `const`, and `volatile`
tokens when parsing the field declaration string. Unfortunately, since
the token `Value` field is a `Span` and not a `String`, the `!=`
operator compares by address, not by value, so we were not correctly
ignoring them. Fix by using `!SequenceEqual` for the comparison.

Also:
- Fix some comments.
- Rename some files to match prior rename of the contained class.